requirements.txt
here in the project i used pycountry library to make sure that the
country which user enter is a real one
i could use list contain names of all cuntries but that is much better

i also used csv library 
